+++ src/main/java/com/solvd/block3/business_logic/Main.java
@@ -2,7 +2,7 @@ package com.solvd.block3.business_logic;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import com.solvd.block3.utilities.Utils;
+import com.solvd.block3.ui.UiUtils;
 
 public class Main 
 {
@@ -10,10 +10,10 @@ public class Main
     public static void main(String[] args)
     {
         LOGGER.info("Welcome to the flight finder. Please input an origin city.");
-        String origin = Utils.makeCity();
+        String origin = UiUtils.makeCity();
 
         LOGGER.info("Now input a destination.");
-        String destination = Utils.makeCity();
+        String destination = UiUtils.makeCity();
 
         //For testing
         LOGGER.info(origin + " " + destination);
