--- src/main/java/com/solvd/block3/ui/SomeClass.java
@@ -1,4 +1,0 @@
-package com.solvd.block3.ui;
-
-public class SomeClass {
-}
